Don't catch OutOfMemoryError
In D2 this is an `Error` rather than an `Exception`, so catching it is
undefined behaviour.  In any case, in an OOM situation it is unlikely
that logging a special message will help much (cf. similar changes in
dhtnode and dlsnode codebases).

The `OutOfMemoryException` alias has in any case been dropped in ocean
v5.x.x, so dropping this check removes one blocker to upgrading.